refactor: rename the OnboardingDao folder
